refactor(hooks): consolidate hook-side banned-terms into one package (U17)
Move banned_terms_{gate,deny,marker}.py into a coherent hooks/scripts/banned_terms/
package (gate/deny/marker submodules) with __init__ re-exporting the public entry
points (handle_banned_terms_pretool, emit_banned_term_deny, resolve_marker,
MarkerVerdict). git-mv preserves history; scan/deny logic unchanged; drop the
per-module bare-name self-alias. Router imports the package entry point.
