+++ hooks/scripts/banned_terms/deny.py
